+++ .github/workflows/javascript-tests.yml
@@ -27,7 +27,7 @@ jobs:
         node-version: ${{ matrix.node-version }}
     - name: npm install, build, and test
       run: |
-        cd javascript
+        cd javascript/linked-list
         npm i
         npm test
       env:
