+++ javascript/link-temp/__tests__/linked-list.test.js
@@ -1,0 +1,10 @@
+'use strict';
+
+// Require our linked list implementation
+const LinkedList = require('../index');
+
+describe('Linked List', () => {
+  it('works', () => {
+    expect(true).toBeTruthy();
+  });
+});
