+++ javascript/linked-list/package.json
