+++ javascript/package.json
@@ -1,5 +1,8 @@
 {
   "name": "data-structures-and-algorithms",
+  "jest": {
+    "testPathIgnorePatterns": ["code-challenges"]
+  },
   "version": "1.0.0",
   "description": "Data Structures, Algorithms, Code Challenges",
   "main": "index.js",
